part3/requirements.txt
django==2.2.9
django-rest-framework
# filters support for django
django-filter
python-social-auth
social-auth-app-django
tornado
websocket-client
